feat(band): add updateBand endpoint to BandController and updateBand method to BandService
The BandController now has a new PATCH endpoint at ':band_id' that allows updating a band. The endpoint expects a 'band_id' parameter and a 'createBand' body object. The 'createBand' object is used to update the band with the specified 'band_id'. The updateBand method in the BandService is responsible for updating the band in the database. If the band with the specified 'band_id' is not found, a NotFoundException is thrown.